+++ R/priors.R
@@ -2,12 +2,10 @@
 #' Priors for Intercept, Main Effects, Interactions
 #'
 #' @description 
-#' The models for rates, probabilities, or means
-#' created with functions [mod_pois()], [mod_binom()],
+#' The models created with functions [mod_pois()], [mod_binom()],
 #' and [mod_norm()] always include an intercept,
 #' and typically include main effects
-#' and interactions formed from
-#' variables in input data.
+#' and interactions formed from variables in input data.
 #' Most models, for instance
 #' include an age effect, and many include
 #' an interaction between age and sex/gender,